# Conflicts:
#	realm/realm-annotations-processor/build.gradle
#	realm/realm-annotations-processor/src/main/java/io/realm/processor/RealmProcessor.java

+++ tools/sync_test_server/ros-testing-server.js
@@ -43,25 +43,28 @@ function handleRequest(request, response) {
 var syncServerChildProcess = null;
 
 // Waits for ROS to be fully initialized.
-function waitForRosToInitialize(attempts, onSuccess, onError) {
+function waitForRosToInitialize(attempts, onSuccess, onError, startSequence) {
     if (attempts == 0) {
         onError("Could not get ROS to start. See Docker log.");
         return;
     }
     http.get("http://0.0.0.0:9080/health", function(res) {
         if (res.statusCode != 200) {
-            winston.info("ROS /health/ returned: " + res.statusCode)
-            waitForRosToInitialize(attempts - 1, onSuccess, onError)
+            winston.warn("ROS /health/ returned: " + res.statusCode)
+            setTimeout(function() {
+                waitForRosToInitialize(attempts - 1, onSuccess, onError, startSequence);
+            }, 500);
         } else {
-            onSuccess();
+            onSuccess(startSequence);
         }
     }).on('error', function(err) {
+        winston.warn("ROS /health/ returned an error: " + err)
         // ROS not accepting any connections yet.
         // Errors like ECONNREFUSED 0.0.0.0:9080 will be reported here.
         // Wait a little before trying again (common startup is ~1 second).
         setTimeout(function() {
-            waitForRosToInitialize(attempts - 1, onSuccess, onError);
-        }, 200);
+            waitForRosToInitialize(attempts - 1, onSuccess, onError, startSequence);
+        }, 500);
     });
 }
 
@@ -112,7 +115,8 @@ function doStartRealmObjectServer(onSuccess, onError) {
                 winston.info(`${data}`);
             });
 
-            waitForRosToInitialize(20, onSuccess, onError);
+            // The interval between every health check is 0.5 second. Give the ROS 30 seconds to get fully initialized.
+            waitForRosToInitialize(60, onSuccess, onError, Date.now());
         }
     });
 }
@@ -121,24 +125,28 @@ function stopRealmObjectServer(onSuccess, onError) {
     if(syncServerChildProcess == null || syncServerChildProcess.killed) {
         onSuccess("No ROS process found or the process has been killed before");
     }
+    if (syncServerChildProcess) {
+        syncServerChildProcess.on('exit', function(code) {
+            winston.info("ROS server stopped due to process being killed. Exit code: " + code);
+            syncServerChildProcess.removeAllListeners('exit');
+            syncServerChildProcess = null;
+            onSuccess();
+        });
 
-    syncServerChildProcess.on('exit', function(code) {
-        winston.info("ROS server stopped due to process being killed. Exit code: " + code);
-        syncServerChildProcess.removeAllListeners('exit');
-        syncServerChildProcess = null;
-        onSuccess();
-    });
-
-    syncServerChildProcess.kill('SIGKILL');
+        syncServerChildProcess.kill('SIGKILL');
+    }
 }
 
 // start sync server
 dispatcher.onGet("/start", function(req, res) {
     winston.info("Attempting to start ROS");
-    startRealmObjectServer(() => {
+    startRealmObjectServer((startSequence) => {
         res.writeHead(200, {'Content-Type': 'text/plain'});
-        res.end('ROS started');
+        let response = `ROS started after ${Date.now() - startSequence} ms`;
+        winston.info(response);
+        res.end(response);
     }, function (err) {
+        winston.error('Starting ROS failed: ' + err);
         res.writeHead(500, {'Content-Type': 'text/plain'});
         res.end('Starting ROS failed: ' + err);
     });
